Script to set up docker
It was created a shell script to set up a mysql container
and change the hosts file to map the ip from the container
to a host. This host is used at spring database configuration.
That makes possible to create e recreate the database as many
as needed without changing the spring configuration.

Signed-off-by: Gustavo Coelho <gutorc@hotmail.com>